Create Student.md
Creating a class as Student. Write a program in Java to display the names and
roll numbers of students. Create an array of 10 students and initialize the array
with user input. Handle ArrayIndexOutOfBoundsExeption, so that any such
problem doesn’t cause illegal termination of the program. Read a character from
the user and display the student names starting with the given character.